More libdpd fixes

+++ psi4/src/psi4/libdpd/file2_mat_wrt.cc
@@ -52,7 +52,7 @@ int DPD::file2_mat_wrt(dpdfile2 *File) {
         coltot = File->params->coltot[h ^ my_irrep];
 
         if (rowtot && coltot)
-            psio_write(File->filenum, File->label, (char *)File->matrix[h][0], rowtot * coltot * sizeof(double),
+            psio_write(File->filenum, File->label, (char *)File->matrix[h][0], static_cast<size_t>(rowtot) * coltot * sizeof(double),
                        irrep_ptr, &next_address);
     }
